renamed and reorganized files for media rather than having seperate movies and series folders
