+++ backend/src/comments/comments.router.js
@@ -1,10 +1,10 @@
 const router = require("express").Router({ mergeParams: true });
-const controller = require("./movieComments.controller");
+const controller = require("./comments.controller");
 const methodNotAllowed = require("../errors/methodNotAllowed");
 
 router
   .route("/")
-  .get(controller.listMovieComments)
+  .get(controller.listComments)
   .post(controller.create)
   .all(methodNotAllowed);
