+++ backend/src/media/media.controller.js
@@ -1,4 +1,4 @@
-const service = require("./movies.service");
+const service = require("./media.service");
 const asyncErrorBoundary = require("../errors/asyncErrorBoundary");
 const validations = require("./validations/validations");
